requirements.txt
python-dotenv
langchain
langchain-openai
streamlit
openai
tiktoken
psycopg2
pgvector
langchain-community
langchain-core
docx2txt